+++ src/config.h
@@ -1,0 +1,14 @@
+#ifndef CONFIG_H
+#define CONFIG_H
+
+#define QUAT_ANIMATION
+#define GTONE_USE_DMP 
+#define SERIAL_PORT Serial
+
+#define GTONE_WIFI
+
+#define VOL_UP 1
+#define VOL_DOWN -1
+
+
+#endif
